updated documentation

+++ src/array/array_get.c
@@ -6,13 +6,13 @@
 /*   By: unite <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/16 23:25:55 by unite             #+#    #+#             */
-/*   Updated: 2020/09/08 22:43:05 by unite            ###   ########.fr       */
+/*   Updated: 2020/09/10 23:23:23 by unite            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "array.h"
 
-char	*array_get(const t_array *array, int index)
+char	*array_get(const t_array *array, size_t index)
 {
 	return (array->arr[index]);
 }
