+++ src/array/array_indexof.c
@@ -6,14 +6,14 @@
 /*   By: unite <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/16 23:24:52 by unite             #+#    #+#             */
-/*   Updated: 2020/09/08 22:54:55 by unite            ###   ########.fr       */
+/*   Updated: 2020/09/10 23:30:13 by unite            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "array.h"
 
-static int	array_indexof_recur(const t_array *array, const char *val,
-								int start, int end)
+static size_t	array_indexof_recur(const t_array *array, const char *val,
+									size_t start, size_t end)
 {
 	int	i;
 	int	cmp;
@@ -30,7 +30,7 @@ static int	array_indexof_recur(const t_array *array, const char *val,
 		return (array_indexof_recur(array, val, start, i));
 }
 
-int			array_indexof(const t_array *array, const char *val)
+size_t			array_indexof(const t_array *array, const char *val)
 {
 	return (array_indexof_recur(array, val, 0, array->size));
 }
