+++ src/array/array_swap.c
@@ -6,13 +6,13 @@
 /*   By: unite <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/17 14:18:30 by unite             #+#    #+#             */
-/*   Updated: 2020/09/08 22:43:46 by unite            ###   ########.fr       */
+/*   Updated: 2020/09/10 23:30:33 by unite            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "array.h"
 
-void	array_swap(t_array *array, int ind1, int ind2)
+void	array_swap(t_array *array, size_t ind1, size_t ind2)
 {
 	char	*content;
